Переносит блок комментариев в отдельный смарт-компонент
